refactor: reorganizar la estructura del proyecto
