--- Frontend-Sistema/src/app/core/models/ProgressTrackerProps.ts
@@ -1,3 +1,0 @@
-export interface ProgressTrackerProps {
-    currentStep: number;
-}
