--- Frontend-Sistema/src/app/core/models/StepDniProps.ts
@@ -1,7 +1,0 @@
-export interface StepDniProps {
-    dni: string;
-    nombres: string;
-    apellidos: string;
-    onNext: () => void;
-    onBack: () => void;
-}
