--- Frontend-Sistema/src/app/core/models/StepEmailProps.ts
@@ -1,5 +1,0 @@
-export interface StepEmailProps {
-    email: string;
-    onNext: () => void;
-    onBack: () => void;
-}
