--- Frontend-Sistema/src/app/core/models/StepProProps.ts
@@ -1,6 +1,0 @@
-export interface StepProProps {
-    formData: any;
-    onChange: (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement>) => void;
-    onNext: () => void;
-    errors?: Record<string, string>;
-}
