--- Frontend-Sistema/src/app/core/models/StepProps.ts
@@ -1,6 +1,0 @@
-export interface StepProps {
-    formData: any;
-    onChange: (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => void;
-    onNext: () => void;
-    errors?: Record<string, string>;
-}
