--- Frontend-Sistema/src/app/core/models/StepSuccessProps.ts
@@ -1,7 +1,0 @@
-export interface StepSuccessProps {
-    formData: {
-        nombres: string;
-        apellidos: string;
-        dni: string;
-    };
-}
